Migrate AccountRestServiceTest to the new test framework. Closes #48148
