+++ test-framework/builders/src/main/java/org/keycloak/testframework/realm/RealmBuilder.java
@@ -613,6 +613,11 @@ public class RealmBuilder extends Builder<RealmRepresentation> {
         return this;
     }
 
+    public RealmBuilder accountTheme(String accountTheme) {
+        rep.setAccountTheme(accountTheme);
+        return this;
+    }
+
     /**
      * Best practice is to use other convenience methods when configuring a realm, but while the framework is under
      * active development there may not be a way to perform all updates required. In these cases this method allows
