+++ tests/base/src/test/java/org/keycloak/tests/account/AccountRestServiceWithUserProfileTest.java
@@ -1,0 +1,368 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.keycloak.tests.account;
+
+import java.io.IOException;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+
+import org.keycloak.admin.client.resource.RealmResource;
+import org.keycloak.events.Details;
+import org.keycloak.events.EventType;
+import org.keycloak.http.simple.SimpleHttpRequest;
+import org.keycloak.models.UserModel;
+import org.keycloak.models.UserModel.RequiredAction;
+import org.keycloak.representations.account.UserRepresentation;
+import org.keycloak.representations.idm.UserProfileAttributeMetadata;
+import org.keycloak.representations.idm.UserProfileMetadata;
+import org.keycloak.representations.userprofile.config.UPAttribute;
+import org.keycloak.representations.userprofile.config.UPAttributePermissions;
+import org.keycloak.representations.userprofile.config.UPConfig;
+import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
+import org.keycloak.testframework.events.EventAssertion;
+import org.keycloak.testsuite.util.userprofile.UserProfileUtil;
+import org.keycloak.userprofile.UserProfileContext;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import static org.keycloak.tests.account.AccountRestServiceTest.assertUserProfileAttributeMetadata;
+import static org.keycloak.tests.account.AccountRestServiceTest.getUserProfileAttributeMetadata;
+import static org.keycloak.testsuite.util.userprofile.UserProfileUtil.PERMISSIONS_ADMIN_EDITABLE;
+import static org.keycloak.testsuite.util.userprofile.UserProfileUtil.PERMISSIONS_ADMIN_ONLY;
+import static org.keycloak.testsuite.util.userprofile.UserProfileUtil.PERMISSIONS_ALL;
+import static org.keycloak.userprofile.config.UPConfigUtils.ROLE_ADMIN;
+import static org.keycloak.userprofile.config.UPConfigUtils.ROLE_USER;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
+
+/**
+ * Test account rest service with custom user profile configurations
+ *
+ * @author Vlastimil Elias <velias@redhat.com>
+ */
+@KeycloakIntegrationTest
+public class AccountRestServiceWithUserProfileTest extends AbstractRestServiceTest {
+
+    private final static String UP_CONFIG_FOR_METADATA = "{\"attributes\": ["
+            + "{\"name\": \"firstName\"," + PERMISSIONS_ALL + ", \"required\": {\"scopes\":[\"profile\"]}, \"displayName\": \"${profile.firstName}\", \"validations\": {\"length\": { \"max\": 255 }}},"
+            + "{\"name\": \"lastName\"," + PERMISSIONS_ALL + ", \"required\": {}, \"displayName\": \"Last name\", \"annotations\": {\"formHintKey\" : \"userEmailFormFieldHint\", \"anotherKey\" : 10, \"yetAnotherKey\" : \"some value\"}},"
+            + "{\"name\": \"attr_with_scope_selector\"," + PERMISSIONS_ALL + ", \"selector\": {\"scopes\": [\"profile\"]}},"
+            + "{\"name\": \"attr_required\"," + PERMISSIONS_ALL + ", \"required\": {}},"
+            + "{\"name\": \"attr_required_by_role\"," + PERMISSIONS_ALL + ", \"required\": {\"roles\" : [\"user\"]}},"
+            + "{\"name\": \"attr_required_by_scope\"," + PERMISSIONS_ALL + ", \"required\": {\"scopes\": [\"profile\"]}},"
+            + "{\"name\": \"attr_not_required_due_to_role\"," + PERMISSIONS_ALL + ", \"required\": {\"roles\" : [\"admin\"]}},"
+            + "{\"name\": \"attr_readonly\"," + PERMISSIONS_ADMIN_EDITABLE + "},"
+            + "{\"name\": \"attr_no_permission\"," + PERMISSIONS_ADMIN_ONLY + "}"
+            + "]}";
+
+    private final static String UP_CONFIG_NO_ACCESS_TO_NAME_FIELDS = "{\"attributes\": ["
+            + "{\"name\": \"firstName\"," + PERMISSIONS_ADMIN_ONLY + ", \"required\": {}, \"displayName\": \"${profile.firstName}\", \"validations\": {\"length\": { \"max\": 255 }}},"
+            + "{\"name\": \"lastName\"," + PERMISSIONS_ADMIN_ONLY + ", \"required\": {}, \"displayName\": \"Last name\", \"annotations\": {\"formHintKey\" : \"userEmailFormFieldHint\", \"anotherKey\" : 10, \"yetAnotherKey\" : \"some value\"}},"
+            + "{\"name\": \"attr_readonly\"," + PERMISSIONS_ADMIN_EDITABLE + "},"
+            + "{\"name\": \"attr_no_permission\"," + PERMISSIONS_ADMIN_ONLY + "}"
+            + "]}";
+
+    private final static String UP_CONFIG_RO_ACCESS_TO_NAME_FIELDS = "{\"attributes\": ["
+            + "{\"name\": \"firstName\"," + PERMISSIONS_ADMIN_EDITABLE + ", \"required\": {}, \"displayName\": \"${profile.firstName}\", \"validations\": {\"length\": { \"max\": 255 }}},"
+            + "{\"name\": \"lastName\"," + PERMISSIONS_ADMIN_EDITABLE + ", \"required\": {}, \"displayName\": \"Last name\", \"annotations\": {\"formHintKey\" : \"userEmailFormFieldHint\", \"anotherKey\" : 10, \"yetAnotherKey\" : \"some value\"}},"
+            + "{\"name\": \"attr_readonly\"," + PERMISSIONS_ADMIN_EDITABLE + "},"
+            + "{\"name\": \"attr_no_permission\"," + PERMISSIONS_ADMIN_ONLY + "}"
+            + "]}";
+
+    private final static String UP_CONFIG_RO_USERNAME_AND_EMAIL = "{\"attributes\": ["
+            + "{\"name\": \"email\"," + PERMISSIONS_ADMIN_EDITABLE + ", \"required\": {}, \"displayName\": \"${email}\", \"annotations\": {\"formHintKey\" : \"userEmailFormFieldHint\", \"anotherKey\" : 10, \"yetAnotherKey\" : \"some value\"}},"
+            + "{\"name\": \"attr_readonly\"," + PERMISSIONS_ADMIN_EDITABLE + "},"
+            + "{\"name\": \"attr_no_permission\"," + PERMISSIONS_ADMIN_ONLY + "}"
+            + "]}";
+
+
+    @Test
+    public void testEditUsernameAllowed() throws IOException {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        setUserProfileConfiguration(UP_CONFIG_FOR_METADATA);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+
+        assertUserProfileAttributeMetadata(user, "username", "${username}", true, false);
+        assertUserProfileAttributeMetadata(user, "email", "${email}", true, false);
+
+        UserProfileAttributeMetadata uam = assertUserProfileAttributeMetadata(user, "firstName", "${profile.firstName}", true, false);
+        Assertions.assertNull(uam.getAnnotations());
+        Map<String, Object> vc = assertValidatorExists(uam, "length");
+        Assertions.assertEquals(255, vc.get("max"));
+
+        uam = assertUserProfileAttributeMetadata(user, "lastName", "Last name", true, false);
+        Assertions.assertNotNull(uam.getAnnotations());
+        Assertions.assertEquals(3, uam.getAnnotations().size());
+        assertAnnotationValue(uam, "formHintKey", "userEmailFormFieldHint");
+        assertAnnotationValue(uam, "anotherKey", 10);
+
+        assertUserProfileAttributeMetadata(user, "attr_with_scope_selector", "attr_with_scope_selector", false, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_required", "attr_required", true, false);
+        assertUserProfileAttributeMetadata(user, "attr_required_by_role", "attr_required_by_role", true, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_required_by_scope", "attr_required_by_scope", true, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_not_required_due_to_role", "attr_not_required_due_to_role", false, false);
+        assertUserProfileAttributeMetadata(user, "attr_readonly", "attr_readonly", false, true);
+
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "attr_no_permission"));
+    }
+
+    @Test
+    public void testGetUserProfileMetadata_NoAccessToNameFields() throws IOException {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        managedRealm.updateWithCleanup(r -> r.editUsernameAllowed(false));
+
+        setUserProfileConfiguration(UP_CONFIG_NO_ACCESS_TO_NAME_FIELDS);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+
+        assertUserProfileAttributeMetadata(user, "username", "${username}", true, true);
+        assertUserProfileAttributeMetadata(user, "email", "${email}", true, false);
+
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "firstName"));
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "lastName"));
+        assertUserProfileAttributeMetadata(user, "attr_readonly", "attr_readonly", false, true);
+
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "attr_no_permission"));
+    }
+
+    @Test
+    public void testUpdateEmailLink() throws Exception {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        managedRealm.cleanup().add(r -> setRequiredActionEnabled(r, RequiredAction.UPDATE_EMAIL, false));
+        managedRealm.updateWithCleanup(r -> r.editUsernameAllowed(false));
+
+        RealmResource realm = managedRealm.admin();
+        setRequiredActionEnabled(realm, RequiredAction.UPDATE_EMAIL, true);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+        assertThat(user.getUserProfileMetadata().getAttributeMetadata(UserModel.EMAIL).getAnnotations().get("kc.required.action.supported"), is(true));
+
+        UPConfig upConfig = realm.users().userProfile().getConfiguration();
+        UPAttribute attribute = upConfig.getAttribute(UserModel.EMAIL);
+        attribute.setPermissions(new UPAttributePermissions(Set.of(ROLE_USER), Set.of(ROLE_ADMIN)));
+        realm.users().userProfile().update(upConfig);
+        user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+        assertThat(user.getUserProfileMetadata().getAttributeMetadata(UserModel.EMAIL).getAnnotations().get("kc.required.action.supported"), is(nullValue()));
+    }
+
+    @Test
+    public void testGetUserProfileMetadata_RoAccessToNameFields() throws IOException {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        managedRealm.updateWithCleanup(r -> r.editUsernameAllowed(false));
+
+        setUserProfileConfiguration(UP_CONFIG_RO_ACCESS_TO_NAME_FIELDS);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+
+        assertUserProfileAttributeMetadata(user, "username", "${username}", true, true);
+        assertUserProfileAttributeMetadata(user, "email", "${email}", true, false);
+
+        assertUserProfileAttributeMetadata(user, "firstName", "${profile.firstName}", true, true);
+        assertUserProfileAttributeMetadata(user, "lastName", "Last name", true, true);
+        assertUserProfileAttributeMetadata(user, "attr_readonly", "attr_readonly", false, true);
+
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "attr_no_permission"));
+    }
+
+    @Test
+    public void testGetUserProfileMetadata_RoAccessToUsernameAndEmail() throws IOException {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        managedRealm.updateWithCleanup(r -> r.editUsernameAllowed(false));
+
+        setUserProfileConfiguration(UP_CONFIG_RO_USERNAME_AND_EMAIL);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+
+        assertUserProfileAttributeMetadata(user, "username", "${username}", true, true);
+        assertUserProfileAttributeMetadata(user, "email", "${email}", true, true);
+
+        assertUserProfileAttributeMetadata(user, "attr_readonly", "attr_readonly", false, true);
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "attr_no_permission"));
+    }
+
+
+    @Test
+    public void testEditUsernameDisallowed() throws IOException {
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+        managedRealm.updateWithCleanup(r -> r.editUsernameAllowed(false));
+
+        setUserProfileConfiguration(UP_CONFIG_FOR_METADATA);
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        Assertions.assertNotNull(user.getUserProfileMetadata());
+
+        assertUserProfileAttributeMetadata(user, "username", "${username}", true, true);
+        assertUserProfileAttributeMetadata(user, "email", "${email}", true, false);
+
+        UserProfileAttributeMetadata uam = assertUserProfileAttributeMetadata(user, "firstName", "${profile.firstName}", true, false);
+        Assertions.assertNull(uam.getAnnotations());
+        Map<String, Object> vc = assertValidatorExists(uam, "length");
+        Assertions.assertEquals(255, vc.get("max"));
+
+        uam = assertUserProfileAttributeMetadata(user, "lastName", "Last name", true, false);
+        Assertions.assertNotNull(uam.getAnnotations());
+        Assertions.assertEquals(3, uam.getAnnotations().size());
+        assertAnnotationValue(uam, "formHintKey", "userEmailFormFieldHint");
+        assertAnnotationValue(uam, "anotherKey", 10);
+
+        assertUserProfileAttributeMetadata(user, "attr_with_scope_selector", "attr_with_scope_selector", false, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_required", "attr_required", true, false);
+        assertUserProfileAttributeMetadata(user, "attr_required_by_role", "attr_required_by_role", true, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_required_by_scope", "attr_required_by_scope", true, false);
+
+        assertUserProfileAttributeMetadata(user, "attr_not_required_due_to_role", "attr_not_required_due_to_role", false, false);
+        assertUserProfileAttributeMetadata(user, "attr_readonly", "attr_readonly", false, true);
+
+        Assertions.assertNull(getUserProfileAttributeMetadata(user, "attr_no_permission"));
+    }
+
+    protected void assertAnnotationValue(UserProfileAttributeMetadata uam, String key, Object value) {
+        Assertions.assertNotNull(uam.getAnnotations(), "Missing annotations for attribute " + uam.getName());
+        Assertions.assertEquals(value, uam.getAnnotations().get(key), "Unexpected value of the " + key + " annotation for attribute " + uam.getName());
+    }
+
+    protected Map<String, Object> assertValidatorExists(UserProfileAttributeMetadata uam, String validatorId) {
+        Assertions.assertNotNull(uam.getValidators(), "Missing validators for attribute " + uam.getName());
+        Assertions.assertTrue(uam.getValidators().containsKey(validatorId), "Missing validator " + validatorId + " for attribute " + uam.getName());
+        return uam.getValidators().get(validatorId);
+    }
+
+    @Test
+    public void testUpdateProfileEventWithAdditionalAttributesAuditing() throws IOException {
+        managedRealm.updateWithCleanup(r -> r.registrationEmailAsUsername(false));
+        registerUserCleanup("test-user@localhost");
+        managedRealm.cleanup().add(r -> UserProfileUtil.setUserProfileConfiguration(r, null));
+
+        setUserProfileConfiguration("{\"attributes\": ["
+                + "{\"name\": \"firstName\"," + PERMISSIONS_ALL + ", \"required\": {}},"
+                + "{\"name\": \"lastName\"," + PERMISSIONS_ALL + ", \"required\": {}},"
+                + "{\"name\": \"attr1\"," + PERMISSIONS_ALL + "},"
+                + "{\"name\": \"attr2\"," + PERMISSIONS_ALL + "}"
+                + "]}");
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        String originalFirstName = user.getFirstName();
+        String originalLastName = user.getLastName();
+        String originalEmail = user.getEmail();
+        user.setAttributes(Optional.ofNullable(user.getAttributes()).orElse(new HashMap<>()));
+
+        user.setEmail("bobby@localhost");
+        user.setFirstName("Homer");
+        user.setLastName("Simpsons");
+        user.getAttributes().put("attr1", Collections.singletonList("val11"));
+        user.getAttributes().put("attr2", Collections.singletonList("val22"));
+
+        events.clear();
+        user = updateAndGet(user, token);
+
+        EventAssertion.assertSuccess(events.poll())
+            .type(EventType.UPDATE_PROFILE)
+            .userId(user.getId())
+            .clientId("account")
+            .details(Details.CONTEXT, UserProfileContext.ACCOUNT.name())
+            .details(Details.PREVIOUS_EMAIL, originalEmail)
+            .details(Details.UPDATED_EMAIL, "bobby@localhost")
+            .details(Details.PREVIOUS_FIRST_NAME, originalFirstName)
+            .details(Details.PREVIOUS_LAST_NAME, originalLastName)
+            .details(Details.UPDATED_FIRST_NAME, "Homer")
+            .details(Details.UPDATED_LAST_NAME, "Simpsons")
+            .details(Details.PREF_UPDATED + "attr2", "val22");
+        Assertions.assertNull(events.poll());
+    }
+
+    @Test
+    public void testManageUserLocaleAttribute() throws IOException {
+        managedRealm.updateWithCleanup(r -> r.internationalizationEnabled(false));
+        registerUserCleanup("test-user@localhost");
+
+        String token = oauth.client("direct-grant", "password").doPasswordGrantRequest("test-user@localhost", "password").getAccessToken();
+        UserRepresentation user = getUser(token);
+        user.setAttributes(Optional.ofNullable(user.getAttributes()).orElse(new HashMap<>()));
+
+        user.getAttributes().put(UserModel.LOCALE, List.of("pt_BR"));
+        user = updateAndGet(user, token);
+        Assertions.assertNull(user.getAttributes());
+
+        managedRealm.updateWithCleanup(r -> r.internationalizationEnabled(true));
+
+        user.singleAttribute(UserModel.LOCALE, "pt_BR");
+        user = updateAndGet(user, token);
+        Assertions.assertEquals("pt_BR", user.getAttributes().get(UserModel.LOCALE).get(0));
+
+        user.getAttributes().remove(UserModel.LOCALE);
+        user = updateAndGet(user, token);
+        Assertions.assertNull(user.getAttributes());
+
+        UserProfileMetadata metadata = user.getUserProfileMetadata();
+
+        Assertions.assertTrue(metadata.getAttributes().stream()
+                .map(UserProfileAttributeMetadata::getName)
+                .filter(UserModel.LOCALE::equals).findAny()
+                .isPresent()
+        );
+    }
+
+    protected void setUserProfileConfiguration(String configuration) {
+        UserProfileUtil.setUserProfileConfiguration(managedRealm.admin(), configuration);
+    }
+
+    protected UserRepresentation getUser(String token) throws IOException {
+        return getUser(token, true);
+    }
+
+    protected UserRepresentation getUser(String token, boolean fetchMetadata) throws IOException {
+        String accountUrl = getAccountUrl(null) + "?userProfileMetadata=" + fetchMetadata;
+        return AccountRestServiceTest.getUser(accountUrl, simpleHttp, token);
+    }
+
+    protected UserRepresentation updateAndGet(UserRepresentation user, String token) throws IOException {
+        SimpleHttpRequest a = simpleHttp.doPost(getAccountUrl(null)).auth(token).json(user);
+        try {
+            Assertions.assertEquals(204, a.asStatus());
+        } catch (AssertionError e) {
+            System.err.println("Error during user update: " + a.asString());
+            throw e;
+        }
+        return getUser(token);
+    }
+}
