+++ tests/custom-providers/src/main/resources/META-INF/keycloak-themes.json
@@ -1,0 +1,10 @@
+{
+  "themes": [
+    {
+      "name": "custom-account-provider",
+      "types": [
+        "account"
+      ]
+    }
+  ]
+}
