+++ tests/custom-providers/src/main/resources/META-INF/services/org.keycloak.services.resource.AccountResourceProviderFactory
@@ -1,0 +1,1 @@
+org.keycloak.tests.providers.theme.CustomAccountResourceProviderFactory
