+++ tests/custom-providers/src/main/resources/theme/custom-account-provider/account/theme.properties
@@ -1,0 +1,1 @@
+accountResourceProvider=ext-custom-account-console
