+++ testsuite/integration-arquillian/tests/base/testsuites/database-suite
@@ -1,4 +1,3 @@
-AccountRestServiceTest
 AuthorizationCodeTest
 AuthorizationTest
 ClientRegistrationTest
