+++ testsuite/integration-arquillian/tests/base/testsuites/jdk-suite
@@ -1,4 +1,3 @@
-AccountRestServiceTest
 AuthorizationCodeTest
 DeployedScriptAuthenticatorTest
 ExportImportTest
